Removed global definition spot for allocators and left delegation to sources
