+++ libpsn00b/libc/malloc.c
@@ -1,12 +1,6 @@
-#include "malloc.h"
+#include <stdlib.h>
 
-#if MALLOC_IMPL == MALLOC_IMPL_TLSF
-#include "tlsf.h"
-
-#elif MALLOC_IMPL == MALLOC_IMPL_AFF
-#include "aff.h"
-
-#elif MALLOC_IMPL == MALLOC_IMPL_CUSTOM
+#if MALLOC_IMPL == MALLOC_IMPL_CUSTOM
 #include <assert.h>
 
 __attribute__((weak))
@@ -27,33 +21,54 @@ void GetHeapUsage(HeapUsage* usage) {
 	abort();
 }
 
-__attribute__((weak))
+__attribute__((weak, hot))
+void free(void* ptr) {
+	_sdk_log("[ERROR] Unimplemented custom allocator handle: void free(void* ptr)\n");
+	abort();
+}
+
+__attribute__((
+	weak,
+	hot,
+	malloc,
+	alloc_size(1)
+#ifdef gnu_version_10
+	, malloc(free, 1)
+#endif
+))
 void* malloc(size_t size) {
 	_sdk_log("[ERROR] Unimplemented custom allocator handle: void* malloc(size_t)\n");
 	abort();
 	return NULL;
 }
 
-__attribute__((weak))
+__attribute__((
+	weak,
+	hot,
+	malloc,
+	alloc_size(1, 2)
+#ifdef gnu_version_10
+	, malloc(free, 1)
+#endif
+))
 void* calloc(size_t num, size_t size) {
 	_sdk_log("[ERROR] Unimplemented custom allocator handle: void* calloc(size_t num, size_t size)\n");
 	abort();
 	return NULL;
 }
 
-__attribute__((weak))
+__attribute__((
+	weak,
+	hot,
+	malloc,
+	alloc_size(2)
+#ifdef gnu_version_10
+	, malloc(free, 1)
+#endif
+))
 void* realloc(void* ptr, size_t size) {
 	_sdk_log("[ERROR] Unimplemented custom allocator handle: void* realloc(void* ptr, size_t size)\n");
 	abort();
 	return NULL;
 }
-
-__attribute__((weak))
-void free(void* ptr) {
-	_sdk_log("[ERROR] Unimplemented custom allocator handle: void free(void* ptr)\n");
-	abort();
-}
-
-#else
-#error Invalid MALLOC_IMPL defined, must be one of (0: custom, 1: AFF, 2: TLSF)
 #endif
