+++ libpsn00b/libc/tlsf.c
@@ -2,6 +2,7 @@
 #include <limits.h>
 #include <stddef.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 
 #include "tlsf.h"
@@ -1305,4 +1306,34 @@ void* tlsf_realloc(tlsf_t tlsf, void* ptr, size_t size)
 	}
 
 	return p;
+}
+
+// ==== API ====
+
+void InitHeap(void* addr, size_t size) {
+	tlsf_init_heap(addr, size);
+}
+
+void TrackHeapUsage(ptrdiff_t alloc_incr) {
+	tlsf_track_heap_usage((void*) 1, alloc_incr);
+}
+
+void GetHeapUsage(HeapUsage* usage) {
+	tlsf_get_heap_usage(usage);
+}
+
+void* malloc(size_t size) {
+	return tlsf_malloc(__tlsf_allocator, size);
+}
+
+void* calloc(size_t num, size_t size) {
+	return tlsf_malloc(__tlsf_allocator, num * size);
+}
+
+void* realloc(void* ptr, size_t size) {
+	return tlsf_realloc(__tlsf_allocator, ptr, size);
+}
+
+void free(void* ptr) {
+	tlsf_free(__tlsf_allocator, ptr);
 }
